TINY-10132: Refactor CaretFormat and FormatUtils modules
